refactor(handler): extract DoCreate from CreateWeb
