+++ pkg/web/handler/create.go
@@ -21,8 +21,6 @@ import (
 	"fmt"
 	"net/http"
 
-	"code.vikunja.io/api/pkg/db"
-	"code.vikunja.io/api/pkg/events"
 	"code.vikunja.io/api/pkg/log"
 	"code.vikunja.io/api/pkg/models"
 	"code.vikunja.io/api/pkg/modules/auth"
@@ -56,44 +54,9 @@ func (c *WebHandler) CreateWeb(ctx *echo.Context) error {
 		return echo.NewHTTPError(http.StatusInternalServerError, "Could not determine the current user.").Wrap(err)
 	}
 
-	// Create the db session
-	s := db.NewSession()
-	defer func() {
-		err = s.Close()
-		if err != nil {
-			log.Errorf("Could not close session: %s", err)
-		}
-	}()
-
-	// Check permissions
-	canCreate, err := currentStruct.CanCreate(s, currentAuth)
-	if err != nil {
-		_ = s.Rollback()
-		events.CleanupPending(s)
+	if err := DoCreate(ctx.Request().Context(), currentStruct, currentAuth); err != nil {
 		return err
 	}
-	if !canCreate {
-		_ = s.Rollback()
-		events.CleanupPending(s)
-		log.Warningf("Tried to create while not having the permissions for it (User: %v)", currentAuth)
-		return echo.NewHTTPError(http.StatusForbidden, "Forbidden")
-	}
-
-	// Create
-	err = currentStruct.Create(s, currentAuth)
-	if err != nil {
-		_ = s.Rollback()
-		events.CleanupPending(s)
-		return err
-	}
-
-	err = s.Commit()
-	if err != nil {
-		events.CleanupPending(s)
-		return err
-	}
-
-	events.DispatchPending(s)
 
 	return ctx.JSON(http.StatusCreated, currentStruct)
 }
